requirements.txt
transformers>=4.20.0
torch>=2.0.0
pillow>=9.0.0
llama-cpp-python>=0.2.0
